+++ CHANGELOG.md
@@ -10,8 +10,12 @@
   - Connect delete certificate to back end api
 - App Detail Pages
   - UI updates of order status button and back button
+- Technical User Management
+  - Enable inactive users list
 - Portal Login Navigation
   - UI updates as per figma
+- UI Cleanups
+  - Removed unused overlays
 
 ### Bugfix
 
@@ -21,6 +25,8 @@
   - fixed help links in configuration overlay
 - Notification
   - fixed help link
+- Release Process
+  - fixed help links in app release process and service release process
 - Unhelpful scroll
   - fixed auto focus issue which resulted in unhelpful scroll in Partner Network and Application Request page
 - App User